[edge linker] fix crash on removing short contours
This implements a more definitive fix to the proposal from
Gabriel Andrade 087f6237d53bdbcff83e1eda906176e9ca91e958

My guess is that the error had to do simply with a wrong way of deleting
list elements within a loop, which invalidated interators.

Compiles, but still has to be tested.